typeck: check_expr_kind -> expr.rs
